requirements.txt
flask
flask-sqlalchemy
flask-login
flask-bcrypt
flask-migrate
psycopg2-binary
python-dotenv
email-validator